requirements.txt
anyio==4.6.2.post1
build==1.2.2.post1
certifi==2024.8.30
colorama==0.4.6
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
packaging==24.1
pyproject_hooks==1.2.0
python-telegram-bot==21.6
sniffio==1.3.1
